Refactored control package. Changed motor power from velocity to raw.
