+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/controls/GamepadControls.java
@@ -1,6 +1,8 @@
-package org.firstinspires.ftc.teamcode.subsystems;
+package org.firstinspires.ftc.teamcode.opmodes.controls;
 
 import com.seattlesolvers.solverslib.gamepad.ToggleButtonReader;
+
+import org.firstinspires.ftc.teamcode.subsystems.Launcher;
 
 public interface GamepadControls {
